neurons/beaver_miner/requirements.txt
llama-cpp-python>=0.2.0
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1
docker>=6.1.0
jinja2>=3.1.0
numpy>=1.24.0
